Bump to Magnesium dependencies
https://docs.opendaylight.org/projects/integration-distribution/en/latest/platform-versions.html
- upgrade pom to use ODL Magnesium target dev dependecies

https://wiki.opendaylight.org/view/Magnesium_platform_upgrade
- force jsr305 version
- disable SpotBugs enforcement in artifacts when needed
- UintXX yangtools migration

https://git.opendaylight.org/gerrit/c/mdsal/+/83960
- migrate class BindingStreamEventWriter
      from org.opendaylight.yangtools.yang.binding package
      to   org.opendaylight.mdsal.binding.dom.codec.api

- unforce maven-bundle-plugin version in api/pom.xml
  and ordmodels/device/pom.xml build sections
- unfoce iana-afn-safi version in ordmodels/device/pom.xml dependencies

TODO:
for next changes to come in the release:
- migrate Junit to yangtools UintXX
- propose fixes to reactivate SpotBugs enforcement
  and get rid of "deprecated marked for removal"

JIRA: TRNSPRTPCE-198

+++ olm/src/test/java/org/opendaylight/transportpce/olm/OlmPowerServiceRpcImplTest.java
@@ -9,7 +9,6 @@
 package org.opendaylight.transportpce.olm;
 
 import com.google.common.util.concurrent.ListenableFuture;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -72,6 +71,7 @@ import org.opendaylight.yang.gen.v1.http.org.openroadm.resource.rev161014.resour
 import org.opendaylight.yang.gen.v1.http.org.openroadm.resource.types.rev161014.ResourceTypeEnum;
 import org.opendaylight.yangtools.yang.common.RpcResult;
 import org.opendaylight.yangtools.yang.common.RpcResultBuilder;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 public class OlmPowerServiceRpcImplTest extends AbstractTest {
 
@@ -146,7 +146,7 @@ public class OlmPowerServiceRpcImplTest extends AbstractTest {
                 .setPmParameterName(new PmParameterNameBuilder()
                 .setExtension("123")
                 .setType(PmNamesEnum.DefectSeconds).build())
-                .setPmParameterValue(new PmDataType(BigInteger.valueOf(1234))).build())
+                .setPmParameterValue(new PmDataType(Uint64.valueOf(1234))).build())
             .build();
         List<Measurements> measurementsList = new ArrayList<Measurements>();
         measurementsList.add(measurements);
